Make signal's faster, benchmark info inside commit Benchmark test: 500'000'000 signals/call's (raw func) Compare [MSVC]: 1) [OLD IMPL] SignalEvent(event) -> 4.5 (avg) sec 2) [NEW IMPL] SignalEvent(event) -> 2.5 (avg) sec (better) 3) std::function -> 0.4 (avg) sec 4) raw call: my_func() 0.018 (avg) sec
Sure, signals are inferior to direct calls, but that's not quite the right test.
Direct calls will of course be faster compared to signals, because signals get context and notify to everyone listeners of type Event.
I think it's a small price to pay for signals, signals simplify your code and allow you to make it more independent instead of getting noodles from dependencies of code.